Correct calculation of average sale prices
Before, it would compute the exponential moving average for each quality band and then the average of those values would be taken as the exponential moving average of the sale price. Now, the sum of individual sale prices divided by the total number of sales is taken as the current average sale price, which is then exponentially averaged.