Temporary implementation of alternative desired wealth algorithm
This algorithm is based on directly using the joint probability distribution of income and wealth from the Wealth and Assets Survey and transforming the propensity to save into a random number characterising the wealth percentile of the household for a given income band. This implementation is temporary hardcoded, it will be stabilised once a decision is made to this regard by comparing results from this algorithm and the original one.