Implements mark-to-market housing wealth and solves bug counting houses
Accounting of households' housing wealth is now based on mark-to-market value looking at the average price of houses of the same quality, instead of the initial sale price as before.

The function returning the number of investment properties has been replaced by a function returning the total number of properties owned. A bug has been solved which was leading to rental properties being counted as owned.