Implement new demographics algorithm
New algorithm relies only on data on the distribution of age. It constructs a distribution of age with monthly bin edges, then computes the expected number of households in each age bin and uses this to implement births and deaths so as to approximately keep this distribution of age.

Aging has been also moved to the Demographics class instead of Household.

Finally, it removes the last parameter related to the previous demographics algorithm.