Solve bug at Demograghics leading to an excess of deaths and births
Randomness at birth process was leading to an excess of agents being created and subsequently destroyed all across the age distribution. This has been solved by making the expected population for each bin an integer and abandoning the randomness corresponding to the previous use of doubles. Also, in order to avoid randomly killing more agents than necessary in higher age bands and then needing to create again these old households, the probability of death is now updated after each implemented death, such that the initial death probability acts as a maximum number of dead agents possible. Note that this will necessarily lead to a slight overpopulation, given that less agents can be killed than expected, but not more.