Temporarily deactivate the amortisation limit for rent prices
This limit was leading to a spurious synchronisation between rental and sales cycles. While the realism of the limit and its consequences on the results of the model should be thought through in the future, deactivating this now should make it easier to understand the current results of the model.