Update initial sale price mechanism and parameters
In particular, this replaces the previous equation, which assumed a log-normal distribution of initial mark-ups, by a new mechanism based on reading the actual mark-up distribution obtained from Zoopla data from a file. The previous dependence on the average time on market of recently sold houses with the same quality is also dropped, as not supported by data. Thus, SALE_MARKUP, SALE_WEIGHT_MONTHS_ON_MARKET and SALE_EPSILON are removed and replaced by DATA_INITIAL_SALE_MARKUP_DIST. The file with the corresponding data is also added. Finally, all references to the housingMarketStats object within the HouseholdBehaviour class are rewired to a local private field.