Update initial rental price setting and reference rental prices
In particular, initial rental prices are now set with a mechanism equivalent to the new initial sale price, that is, a random mark-up over previous average rental prices for houses of the same quality is drawn from a real distribution of rental mark-ups, which is read from data. The corresponding data file is also hereby added. Regarding reference rental prices, these are now set based on a log-normal distribution of rental prices, with parameters fit with English Housing Survey data for 2012-2013.