Update both tenancy length parameters and their source
The definition of the parameters is slightly changed, such that now they set the minimum and the maximum of the tenancy length, making it more coherent with the use of a uniform (discrete) distribution. As a consequence, the average tenancy length is now introduced as a derived parameter. Finally, the specific supporting document is also uploaded.