Cap desired consumption to a given fraction of gross annual income
Note that the value chosen now for this parameter is not yet properly calibrated. The value of the market average price decay is also updated to 1/2, such that all households value equally rental income and capital gains. Also, the Bank's rate spread response to demand for credit is temporarily changed to a very large value so as to keep mortgage rates basically fixed (evolving too slowly). This value should be updated later on.