Update renter's desired expenditure mechanism
The desired rental price for households to bid on the rental market is now decided in a similar way as the buyer's desired expenditure (which is itself inspired by Axtell's paper): it is modelled as a nonlinear function of the household's annual income, capped to it's available monthly net income after essential consumption. For the calibration of the new associated parameters, BoE's NMG survey data for 2016 has been used.