Update CreditSupply and CoreIndicators
Among other things, DescriptiveStatistics objects are replaced by ArrayLists at CreditSupply, the computation of the mean above the median is taken from an implementation of a UnivariateStatistic back to CoreIndicators and simplified, thus the MeanAboveMedian class is removed. Averaging over a user-defined number of months, with a rolling window, is introduced for some core indicators.