Prevent households considering too large bid price
Exploratory bid prices at decideRentOrPurchase() are now capped to the
average price of the maximum quality band, in order to prevent
households considering the huge mortgage costs corresponding to those
exaggeratedly large bids. Instead, the mortgage costs of more realistic
maximum sale prices are considered.

Furthermore, a mechanism to prevent negative and too small price updates
is introduced both for properties failing to be sold and to be rented
out.