Simplify and improve makeMonthlyPayment method at MortgageAgreement
The method is improved in the sense that any principal left is now paid off at the final month of the mortgage, instead of a month later (for which no interest was paid). Simplification is due to the removal of the isActive variable, not needed anymore. In this way, temporary negative principals and monthly payments are avoided, since final settlement of dues takes place joined to last monthly payment.