New initial free and random allocation of houses
This removes a great deal of the initial instability, avoiding a too large spike in buyer, sellers, sales, rental prices and interest rates, as well as an exaggerated drop in sale prices. While a burn-out period is still required, since a random house allocation is clearly not realistic, this new mechanism leads to a much smoother initialisation, with most variables relaxing to their trend values fairly quickly.

Parts of the old mechanism (construction sector selling the houses on the market) are kept within the code, as they might be useful in the future, when a more developed construction sector is implemented.