Update interest rate dynamics and parameters
In particular, interest rate changes are now a function of credit supply
changes, as opposed to a function of the distance to the target credit
supply, as it was before. The parameter of this equation is now
inverted, thus being a differential interest rate over a differential
credit demand/supply change. The previous target credit supply is now
replaced by an initial credit supply, and it also becomes less relevant
for the dynamics of the model.