Apply soft LTI limit at approval in principle letter stage
This modifies the implementation of the soft limit LTI policy so that the bank can choose which LTI to use at the approval in principle letter stage. In particular, 1) households now store the LTI limit assigned to them at the approval in principle letter stage and use it at the actual mortgage request stage and 2) the bank assigns LTI limits at the approval in principle stage trying to match the corresponding target at the actual mortgage stage. Thus, the fraction of actual mortgages over the soft limit can eventually, and temporarily, overshoot its target. This solves the problem that FTB households in the previous method would always follow the soft limit, with no fraction over it.